# src/main/java/com/example/store/categoria/dto/CategoriaRequest.java
package com.example.store.categoria.dto;

import com.fasterxml.jackson.annotation.JsonAlias;
import jakarta.validation.constraints.NotBlank;

public class CategoriaRequest {

    @NotBlank(message = "El nombre es obligatorio")
    private String nombre;

    @JsonAlias("parentId")
    private Long padreId;

    public String getNombre() { return nombre; }
    public void setNombre(String nombre) { this.nombre = nombre; }
    public Long getPadreId() { return padreId; }
    public void setPadreId(Long padreId) { this.padreId = padreId; }
}