# src/main/java/com/example/store/orden/dto/CambiarEstadoRequest.java
package com.example.store.orden.dto;

import jakarta.validation.constraints.NotBlank;

public class CambiarEstadoRequest {

    @NotBlank(message = "El estado es obligatorio")
    private String estado;

    public String getEstado() { return estado; }
    public void setEstado(String estado) { this.estado = estado; }
}